refactor(website): extract shared deriveMarkdownHref helper
Deduplicate the permalink→`.md` normalization between the DocBreadcrumbs
and DocItem/Content swizzles by moving it into a shared
`website/src/theme/docUtils.ts` module. Both swizzles must point at the
same `.md` URL, and the inline copy in DocItem/Content already carried a
"match the breadcrumbs swizzle" comment — extraction removes the drift
risk.
